+++ src/app-settings/app-settings.ts
@@ -5,24 +5,12 @@ export const appSettings = {
   ADS_OPERATOR_URL: environment.production ? 'https://operator.adshares.net' : 'https://operator1.e11.click',
   // milliseconds of transitioning effect between views
   ROUTER_TRANSITION_DURATION: 200,
-  // maximal size of image ad in bytes
-  MAX_AD_SIZE_IMAGE: 512 * 1024,
-  // maximal size of html ad in bytes
-  MAX_AD_SIZE_HTML: 512 * 1024,
-  // maximal size of video ad in bytes
-  MAX_AD_SIZE_VIDEO: 1024 * 1024,
-  // maximal size of model ad in bytes
-  MAX_AD_SIZE_MODEL: 1024 * 1024,
   // seconds of how long remembered user auth token will be stored
   REMEMBER_USER_EXPIRATION_SECONDS: 259200,
   // seconds of how long not remembered user auth token will be stored
   AUTH_TOKEN_EXPIRATION_SECONDS: 900,
   // time after which advertisers'/publishers' charts and tables will be automatically refreshed
   AUTOMATIC_REFRESH_INTERVAL: 60000,
-  // update notification time in milliseconds
-  UPDATE_NOTIFICATION_MILLISECONDS_INTERVAL: 60000,
-  // dismiss push notification time in milliseconds
-  DISMISS_PUSH_NOTIFICATION_INTERVAL: 10000,
   // withdraw amounts values
   WITHDRAWAL_AMOUNTS: [100, 200, 300, 400, 500, 600, 700, 800, 900, 1000],
   // privacy policy link