+++ src/app/common/components/settings-menu-item/settings-menu-item.component.ts
@@ -7,8 +7,4 @@ import { Component, Input } from '@angular/core';
 })
 export class SettingsMenuItemComponent {
   @Input() menuItem;
-
-  constructor(
-  ) {
-  }
 }